eliminate!: Remove now deprecated `EngineReportingExtension`.
The plugin implementation brought in de7ba72b50ff5 supersedes the need for
this implementation!
