+++ packages/apollo-engine-reporting/src/plugin.ts
@@ -1,18 +1,10 @@
 import {
   GraphQLRequestContext,
-  WithRequired,
   Logger,
-  SchemaHash,
   InvalidGraphQLRequestError,
 } from 'apollo-server-types';
-import { Request, Headers } from 'apollo-server-env';
-import {
-  GraphQLResolveInfo,
-  DocumentNode,
-  ExecutionArgs,
-  GraphQLError,
-} from 'graphql';
-import { GraphQLExtension, EndHandler } from 'graphql-extensions';
+import { Headers } from 'apollo-server-env';
+import { GraphQLError } from 'graphql';
 import { Trace } from 'apollo-engine-reporting-protobuf';
 
 import {
@@ -34,187 +26,6 @@ type Mutable<T> = { -readonly [P in keyof T]: T[P] };
 const clientNameHeaderKey = 'apollographql-client-name';
 const clientReferenceIdHeaderKey = 'apollographql-client-reference-id';
 const clientVersionHeaderKey = 'apollographql-client-version';
-
-// EngineReportingExtension is the per-request GraphQLExtension which creates a
-// trace (in protobuf Trace format) for a single request. When the request is
-// done, it passes the Trace back to its associated EngineReportingAgent via the
-// addTrace callback in its constructor. This class isn't for direct use; its
-// constructor is a private API for communicating with EngineReportingAgent.
-// Its public methods all implement the GraphQLExtension interface.
-export class EngineReportingExtension<TContext = any>
-  implements GraphQLExtension<TContext> {
-  private logger: Logger = console;
-  private treeBuilder: EngineReportingTreeBuilder;
-  private explicitOperationName?: string | null;
-  private queryString?: string;
-  private documentAST?: DocumentNode;
-  private options: EngineReportingOptions<TContext>;
-  private addTrace: (args: AddTraceArgs) => Promise<void>;
-  private generateClientInfo: GenerateClientInfo<TContext>;
-
-  public constructor(
-    options: EngineReportingOptions<TContext>,
-    addTrace: (args: AddTraceArgs) => Promise<void>,
-    private schemaHash: SchemaHash,
-  ) {
-    this.options = {
-      ...options,
-    };
-    if (options.logger) this.logger = options.logger;
-    this.addTrace = addTrace;
-    this.generateClientInfo =
-      options.generateClientInfo || defaultGenerateClientInfo;
-
-    this.treeBuilder = new EngineReportingTreeBuilder({
-      rewriteError: options.rewriteError,
-      logger: this.logger,
-    });
-  }
-
-  public requestDidStart(o: {
-    request: Request;
-    queryString?: string;
-    parsedQuery?: DocumentNode;
-    variables?: Record<string, any>;
-    context: TContext;
-    extensions?: Record<string, any>;
-    requestContext: WithRequired<
-      GraphQLRequestContext<TContext>,
-      'metrics' | 'queryHash'
-    >;
-  }): EndHandler {
-    this.treeBuilder.startTiming();
-    o.requestContext.metrics.startHrTime = this.treeBuilder.startHrTime;
-
-    // Generally, we'll get queryString here and not parsedQuery; we only get
-    // parsedQuery if you're using an OperationStore. In normal cases we'll get
-    // our documentAST in the execution callback after it is parsed.
-    const queryHash = o.requestContext.queryHash;
-    this.queryString = o.queryString;
-    this.documentAST = o.parsedQuery;
-
-    this.treeBuilder.trace.http = new Trace.HTTP({
-      method:
-        Trace.HTTP.Method[o.request.method as keyof typeof Trace.HTTP.Method] ||
-        Trace.HTTP.Method.UNKNOWN,
-      // Host and path are not used anywhere on the backend, so let's not bother
-      // trying to parse request.url to get them, which is a potential
-      // source of bugs because integrations have different behavior here.
-      // On Node's HTTP module, request.url only includes the path
-      // (see https://nodejs.org/api/http.html#http_message_url)
-      // The same is true on Lambda (where we pass event.path)
-      // But on environments like Cloudflare we do get a complete URL.
-      host: null,
-      path: null,
-    });
-
-    if (this.options.sendHeaders) {
-      makeHTTPRequestHeaders(
-        this.treeBuilder.trace.http,
-        o.request.headers,
-        this.options.sendHeaders,
-      );
-
-      if (o.requestContext.metrics.persistedQueryHit) {
-        this.treeBuilder.trace.persistedQueryHit = true;
-      }
-      if (o.requestContext.metrics.persistedQueryRegister) {
-        this.treeBuilder.trace.persistedQueryRegister = true;
-      }
-    }
-
-    if (o.variables) {
-      this.treeBuilder.trace.details = makeTraceDetails(
-        o.variables,
-        this.options.sendVariableValues,
-        o.queryString,
-      );
-    }
-
-    const clientInfo = this.generateClientInfo(o.requestContext);
-    if (clientInfo) {
-      // While clientAddress could be a part of the protobuf, we'll ignore it for
-      // now, since the backend does not group by it and Engine frontend will not
-      // support it in the short term
-      const { clientName, clientVersion, clientReferenceId } = clientInfo;
-      // the backend makes the choice of mapping clientName => clientReferenceId if
-      // no custom reference id is provided
-      this.treeBuilder.trace.clientVersion = clientVersion || '';
-      this.treeBuilder.trace.clientReferenceId = clientReferenceId || '';
-      this.treeBuilder.trace.clientName = clientName || '';
-    }
-
-    return () => {
-      this.treeBuilder.stopTiming();
-
-      this.treeBuilder.trace.fullQueryCacheHit = !!o.requestContext.metrics
-        .responseCacheHit;
-      this.treeBuilder.trace.forbiddenOperation = !!o.requestContext.metrics
-        .forbiddenOperation;
-      this.treeBuilder.trace.registeredOperation = !!o.requestContext.metrics
-        .registeredOperation;
-
-      // If the user did not explicitly specify an operation name (which we
-      // would have saved in `executionDidStart`), but the request pipeline made
-      // it far enough to figure out what the operation name must be and store
-      // it on requestContext.operationName, use that name.  (Note that this
-      // depends on the assumption that the RequestContext passed to
-      // requestDidStart, which does not yet have operationName, will be mutated
-      // to add operationName later.)
-      const operationName =
-        this.explicitOperationName || o.requestContext.operationName || '';
-      const documentAST = this.documentAST || o.requestContext.document;
-
-      // If this was a federated operation and we're the gateway, add the query plan
-      // to the trace.
-      if (o.requestContext.metrics.queryPlanTrace) {
-        this.treeBuilder.trace.queryPlan =
-          o.requestContext.metrics.queryPlanTrace;
-      }
-
-      this.addTrace({
-        operationName,
-        queryHash,
-        documentAST,
-        queryString: this.queryString || '',
-        trace: this.treeBuilder.trace,
-        schemaHash: this.schemaHash,
-      });
-    };
-  }
-
-  public executionDidStart(o: { executionArgs: ExecutionArgs }) {
-    // If the operationName is explicitly provided, save it. Note: this is the
-    // operationName provided by the user. It might be empty if they're relying on
-    // the "just use the only operation I sent" behavior, even if that operation
-    // has a name.
-    //
-    // It's possible that execution is about to fail because this operation
-    // isn't actually in the document. We want to know the name in that case
-    // too, which is why it's important that we save the name now, and not just
-    // rely on requestContext.operationName (which will be null in this case).
-    if (o.executionArgs.operationName) {
-      this.explicitOperationName = o.executionArgs.operationName;
-    }
-    this.documentAST = o.executionArgs.document;
-  }
-
-  public willResolveField(
-    _source: any,
-    _args: { [argName: string]: any },
-    _context: TContext,
-    info: GraphQLResolveInfo,
-  ): ((error: Error | null, result: any) => void) | void {
-    return this.treeBuilder.willResolveField(info);
-    // We could save the error into the trace during the end handler, but it
-    // won't have all the information that graphql-js adds to it later, like
-    // 'locations'.
-  }
-
-  public didEncounterErrors(errors: GraphQLError[]) {
-    this.treeBuilder.didEncounterErrors(errors);
-  }
-}
 
 // This plugin is instantiated once at server start-up. Each request that the
 // server processes will invoke the `requestDidStart` method which will produce
