+++ packages/apollo-engine-reporting/tsconfig.json
@@ -7,7 +7,6 @@
   "include": ["src/**/*"],
   "exclude": ["**/__tests__", "**/__mocks__"],
   "references": [
-    { "path": "../graphql-extensions" },
     { "path": "../apollo-server-errors" },
     { "path": "../apollo-server-types" },
     { "path": "../apollo-server-plugin-base" },
